backend/requirements.txt
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.2
django-cors-headers==3.0.2
django-debug-toolbar==1.11
django-extensions==2.1.9
django-rest-auth==0.9.5
django-rest-passwordreset==1.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pip==19.1.1
prompt-toolkit==2.0.9
psycopg2==2.8.3
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
setuptools==41.0.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
